+++ crayg/scene/Scene.h
@@ -10,6 +10,7 @@
 #include "intersectors/Intersectable.h"
 #include "Sphere.h"
 #include "Light.h"
+#include "Camera.h"
 
 // todo add camera member and Rendersettings
 class Scene{
@@ -23,5 +24,8 @@ public:
     std::vector<std::shared_ptr<SceneObject>> objects;
     std::vector<std::shared_ptr<Light>> lights;
 
+    Camera* camera= nullptr;
+
+    virtual ~Scene();
 };
 #endif //CRAYG_SCENE_H